Adding confirmation dialog as Hook so I can reuse later.

+++ components/data-table.tsx
@@ -24,6 +24,8 @@ import {
 import {Button} from "@/components/ui/button";
 import { Input } from "@/components/ui/input"
 
+import { useConfirm } from "@/hooks/use-confirm";
+
 import React from "react";
 import {Trash} from "lucide-react";
 
@@ -42,6 +44,11 @@ export function DataTable<TData, TValue>({
                                              onDelete,
                                              disabled,
                                          }: DataTableProps<TData, TValue>) {
+
+    const [ConfirmationDialog, confirm] = useConfirm(
+        "Are you sure?",
+        "you are about to perform a bulk delete."
+    )
 
     const [sorting, setSorting] = React.useState<SortingState>([])
     const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
@@ -69,6 +76,7 @@ export function DataTable<TData, TValue>({
 
     return (
         <div>
+            <ConfirmationDialog/>
             <div className="flex items-center py-4">
                 <Input
                     placeholder={`Filter ${filterKey}...`}
@@ -84,9 +92,12 @@ export function DataTable<TData, TValue>({
                         size="sm"
                         variant="outline"
                         className="ml-auto font-normal text-xs"
-                        onClick={() =>{
-                            onDelete(table.getFilteredSelectedRowModel().rows)
-                            table.resetRowSelection();
+                        onClick={async () =>{
+                            const ok = await confirm();
+                            if(ok){
+                                onDelete(table.getFilteredSelectedRowModel().rows)
+                                table.resetRowSelection();
+                            }
                         }}
                     >
                         <Trash className="size-4 mr-2"/>
